feat: add fields to fields folder
